safari-600.7-branch/LayoutTests/fast/inspector-support/cssURLQuotes-expected.txt

ALERT: url(file:///unquoted) (URL should not be quoted)
ALERT: url(file:///noQuotesNeeded) (URL should not be quoted)
ALERT: url('file:///should(Quote)') (URL should be quoted)
